+++ installers/raspberry.sh
@@ -113,6 +113,34 @@ else
 	exit;
 fi
 
+# Check if plymouth is installed (default with PIXEL desktop environment), then install custom splashscreen.
+echo -e "\e[96mCheck plymouth installation ...\e[0m"
+if command_exists plymouth; then
+    THEME_DIR="/usr/share/plymouth/themes"
+    echo -e "\e[90mSplashscreen: Checking themes directory.\e[0m"
+    if [ -d $THEME_DIR ]; then
+        echo -e "\e[90mSplashscreen: Create theme directory if not exists.\e[0m"
+        if [ ! -d $THEME_DIR/MagicMirror ]; then
+            sudo mkdir $THEME_DIR/MagicMirror
+        fi
+
+        if sudo cp ~/MagicMirror/splashscreen/splash.png $THEME_DIR/MagicMirror/splash.png && sudo cp ~/MagicMirror/splashscreen/MagicMirror.plymouth $THEME_DIR/MagicMirror/MagicMirror.plymouth && sudo cp ~/MagicMirror/splashscreen/MagicMirror.script $THEME_DIR/MagicMirror/MagicMirror.script; then
+            echo -e "\e[90mSplashscreen: Theme copied successfully.\e[0m"
+            if sudo plymouth-set-default-theme -R MagicMirror; then
+                echo -e "\e[92mSplashscreen: Changed theme to MagicMirror successfully.\e[0m"
+            else
+                echo -e "\e[91mSplashscreen: Couldn't change theme to MagicMirror!\e[0m"
+            fi
+        else
+            echo -e "\e[91mSplashscreen: Copying theme failed!\e[0m"
+        fi
+    else
+        echo -e "\e[91mSplashscreen: Themes folder doesn't exist!\e[0m"
+    fi
+else
+	echo -e "\e[93mplymouth is not installed.\e[0m";
+fi
+
 echo " "
 echo -e "\e[92mWe're ready! Run \e[1m\e[97mDISPLAY=:0 npm start\e[0m\e[92m from the ~/MagicMirror directory to start your MagicMirror.\e[0m"
 echo " "